% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltt_func.R
\name{get_ltt_func}
\alias{get_ltt_func}
\title{Get the lineage-through-time function of a tree}
\usage{
get_ltt_func(phylo)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).}
}
\value{
a stepwise function of time that returns the number of lineages
in the tree at any point in time
}
\description{
Builds the lineage-through-time table from the tree, and converts it
to a stepwise function of time with \code{\link[stats:stepfun]{stats::stepfun()}}
}
\author{
Théo Pannetier
}
