% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pars_doc.R
\name{pars_doc}
\alias{pars_doc}
\title{Common parameters documentation}
\usage{
pars_doc(phylo)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).}
}
\description{
This function does nothing, it is used for automated documentation of
common parameters of the package
}
\author{
Théo Pannetier
}
