% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ltt.R
\name{plot_ltt}
\alias{plot_ltt}
\title{Lineage-through-time plot for a tree}
\usage{
plot_ltt(phylo)
}
\arguments{
\item{phylo}{an object of the class \code{phylo} as introduced in
\link[ape:read.tree]{ape}. The phylogeny must start with the crown node
(not stem), and be binary (no hard polytomies).}
}
\description{
Uses \code{ggplot} to plot the LTT of a tree
}
\author{
Théo Pannetier
}
